requirements.txt
django
djangorestframework
django-cors-headers
scipy
librosa
